fix billing_cycle wise

+++ database/migrations/2026_09_18_105600_add_fields_to_email_templates_table.php
@@ -12,17 +12,21 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('email_templates', function (Blueprint $table) {
-            $table->enum('status', ['active', 'inactive'])->default('active')->after('images');
-            
-            $table->unsignedBigInteger('created_by')->nullable()->after('status');
-            $table->unsignedBigInteger('updated_by')->nullable()->after('created_by');
-            $table->unsignedBigInteger('deleted_by')->nullable()->after('updated_by');
-            
-            $table->foreign('created_by')->references('id')->on('users')->onDelete('set null');
-            $table->foreign('updated_by')->references('id')->on('users')->onDelete('set null');
-            $table->foreign('deleted_by')->references('id')->on('users')->onDelete('set null');
-            
-            $table->softDeletes()->after('updated_at');
+            if (!Schema::hasColumn('email_templates', 'status')) {
+                $table->enum('status', ['active', 'inactive'])->default('active')->after('images');
+            }
+            if (!Schema::hasColumn('email_templates', 'created_by')) {
+                $table->unsignedBigInteger('created_by')->nullable()->after('status');
+                $table->unsignedBigInteger('updated_by')->nullable()->after('created_by');
+                $table->unsignedBigInteger('deleted_by')->nullable()->after('updated_by');
+                
+                $table->foreign('created_by')->references('id')->on('users')->onDelete('set null');
+                $table->foreign('updated_by')->references('id')->on('users')->onDelete('set null');
+                $table->foreign('deleted_by')->references('id')->on('users')->onDelete('set null');
+            }
+            if (!Schema::hasColumn('email_templates', 'deleted_at')) {
+                $table->softDeletes()->after('updated_at');
+            }
         });
     }
